requirements.txt
telethon==1.36.0
schedule==1.2.0
python-dotenv==1.0.0
pytz==2024.1
